Change how I'm handling the slider tap action
FIX: No longer triggers the tap_action when sliding!
BUG: does not trigger tap_action unless slider is disabled
It is supposed to trigger tap_action at all times.

The goal is that this change will eventually result in "Act as a toggle (disable
sliding)" being removed. A toggle for disabling slider will remain, but
it will no longer add a tap action for toggle when enabled, as that will be attached to the
slider anyway (and can be defined as 'none' action if desired).